H2 and MySql datasources
Maven configuration provides possibility to set profiles for development and production environment  (i.e. mvn install -Pproduction). It is needed to specify profile while building. It is necessary to to set JAVA_OPTS on application server. For Wildfly bin/standalone.conf

Problems:
* Tests doesn't recognize datasource
* Error page redirection while login on production environment. Though no error

+++ src/main/java/ru/cadmy/finance/service/UserService.java
@@ -14,13 +14,13 @@ public interface UserService {
 
     public List<User> getUserList();
 
-    public void removeUser(Integer id);
+    public void removeUser(Long id);
 
     public User getCurrentUser();
 
     public String getCurrentUsername();
 
-    public User getUserById(Integer userId);
+    public User getUserById(Long userId);
 
     public User getUserByUsername(String username);
 }
